Dev

See merge request documentale/sprint-flows!630

+++ src/main/java/it/cnr/si/service/DraftService.java
@@ -89,7 +89,7 @@ public class DraftService {
      */
     public void deleteDraftByTaskIdAndUsername(Long taskId, String username) {
         Draft draft = draftRepository.getDraftByTaskIdAndUsername(taskId, username);
-
-        draftRepository.delete(draft.getId());
+        if(draft != null)
+            draftRepository.delete(draft.getId());
     }
 }
